% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_formulas.R
\docType{data}
\name{baseline_formulas}
\alias{baseline_formulas}
\title{Lists of baselines formulas we are using.  This is a list of lists with three things}
\format{
A list with two elements: a list of formulas and a data frame
}
\usage{
data(mauna_loa)
}
\description{
\itemize{
  \item formulas. A list of the name of the formula, with its regression formula, usually with time as the explanatory (x) variable and value is the response (y) variable.
  \item formula_info. A data frame with two columns: the first is the model name (consistent with formuals) and the second is the degrees of freedom.
}
}
\keyword{datasets}
