% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorado.R
\docType{data}
\name{colorado}
\alias{colorado}
\title{Net carbon uptake from the Niwot Ridge FLUXNET site}
\format{A data frame with 5844 rows and 4 variables}
\source{
Zobitz, John. 2013. “Forest Carbon Uptake and the Fundamental Theorem of Calculus.” The College Mathematics Journal 44 (5): 421–24. https://doi.org/10.4169/college.math.j.44.5.421.
}
\usage{
data(colorado)
}
\description{
\itemize{
  \item date. POSIX date, but measurement started in 1998. We should have one measurement per day.
  \item time. The decimal representation of the date.
  \item product. "cumNEE" cumulative net carbon uptake, expressed in gC m^-2
  \item value. value of measurement in previous column
}
}
\keyword{datasets}
