% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_results.R
\docType{data}
\name{data_results}
\alias{data_results}
\title{A list of fitted results for the colorado, nicarauga, and mauna_loa datasets}
\format{A list with 3 elements. 1) models --> linear model fit structures for the data fit, 2) aic --> AIC calculates for each of the fits, 3) fitted --> fitted values of our results.  See the file data-fit.R in the fit-process folder}
\usage{
data(data_results)
}
\description{
A list of fitted results for the colorado, nicarauga, and mauna_loa datasets
}
\keyword{datasets}
