% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{fit_stats}
\alias{fit_stats}
\title{Compute model data statistics.}
\usage{
fit_stats(model, observations, n_params)
}
\arguments{
\item{model}{The modeled values}

\item{observations}{The observed values.  Must be the same size as model}

\item{n_params}{the number of parameters}
}
\value{
A data frame AIC, R2, and centered root mean square values
}
\description{
\code{fit_stats} Computes the following model - data statistics
\itemize{
  \item AIC
  \item R squared value
  \item Centered root mean square value
  }
}
