% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPP_data.R
\docType{data}
\name{GPP_data}
\alias{GPP_data}
\title{Gross Primary Productivity for 3 MODIS pixels.}
\format{A data frame with 1561 rows and 5 variables}
\source{
Running, S., Mu, Q., Zhao, M. (2015). MOD17A2H MODIS/Terra Gross Primary Productivity 8-Day L4 Global 500m SIN Grid V006. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD17A2H.006. Accessed January 4, 2019.

Running, S., Mu, Q., Zhao, M. (2015). MYD17A2H MODIS/Aqua Gross Primary Productivity 8-Day L4 Global 500m SIN Grid V006. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD17A2H.006. Accessed January 4, 2019.
}
\usage{
data(GPP_data)
}
\description{
\itemize{
  \item site. Fluxnet site abbreviation
  \item date. POSIX date.
  \item time. The decimal representation of the date.
  \item product. "GPP" Gross primary productivity
  \item value. value of measurement in previous column
}
}
\keyword{datasets}
