% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_fit_process.R
\name{joint_fit_process}
\alias{joint_fit_process}
\title{Compute a baseline and periodic fit to data.}
\usage{
joint_fit_process(data, model_stats_filename)
}
\arguments{
\item{data}{a data frame with three columns: date (a POSIX date), time (the decimal time) and the value}

\item{model_stats_filename}{where you want the model stats file to be stored}
}
\value{
A data frame of fitted values than we can use for plotting
}
\description{
\code{joint_fit_process} Is an overarching analysis function that does a lot of heavy lifting:
\itemize{
  \item Computes a periodic and trigonometric fit (using periodic_fit_process and trigonometric_fit_process)
  \item Analyzes the two fits and excludes out any fits where the the degrees of freedom is less than expected
  \item Writes a data table of AIC and BIC values for analysis.
  \item Returns fitted values for the periodic and trigonometric fits for plotting.
}Returns a list of fitted and predicted values of periodically repeating data.  This works over a list of data sites efficiently processing multiple locations
}
\examples{

# This really is an overarching function that does the master fitting, since we are doing a lot of things repeatedly this seemed to be the easiest to avoid errors!

}
