% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_taylor_process.R
\name{joint_taylor_process}
\alias{joint_taylor_process}
\title{Compute a baseline and periodic fit to data.}
\usage{
joint_taylor_process(data, formulas)
}
\arguments{
\item{data}{a listed data frame with three columns: date (a POSIX date), time (the decimal time) and the value}

\item{formulas}{a list of formulas we want to use in the fitting}
}
\value{
A data frame of fitted values than we can use for plotting
}
\description{
\code{joint_taylor_process} Is an overarching analysis function that does a lot of heavy lifting:
\itemize{
  \item Computes a periodic and trigonometric fit (using periodic_fit_process and trigonometric_fit_process)
  \item Analyzes the two fits and excludes out any fits where the the degrees of freedom is less than expected
  \item Identifies the necessary information to plot on a Taylor diagram
}Returns a list of fitted and predicted values of periodically repeating data.  This works over a list of data sites efficiently processing multiple locations
}
\examples{

# This really is an overarching function that does the master fitting, since we are doing a lot of things repeatedly this seemed to be the easiest to avoid errors!

}
