% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mauna_loa.R
\docType{data}
\name{mauna_loa}
\alias{mauna_loa}
\title{Mauna Loa carbon dioxide data}
\format{A data frame with 722 rows and 2 variables}
\source{
\url{https://www.esrl.noaa.gov/gmd/ccgg/trends/data.html}
I downloaded the data and then modified it in Excel to strip trailing white space, makes it easier to read in.
}
\usage{
data(mauna_loa)
}
\description{
\itemize{
  \item time. POSIX date, but measurement started in 1958
  \item co2 expressed as a mole fraction in dry air, micromol/mol, abbreviated as ppm
}
}
\keyword{datasets}
