% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd43A3_results.R
\docType{data}
\name{mcd43A3_results}
\alias{mcd43A3_results}
\title{A list of fitted results for the black sky albedo (mcd43A3) data products}
\format{A list with 3 elements. 1) models --> linear model fit structures for the data fit, 2) aic --> AIC calculates for each of the fits, 3) fitted --> fitted values of our results.  See the file albedo-fit.R in the fit-process folder}
\usage{
data(mcd43A3_results)
}
\description{
A list of fitted results for the black sky albedo (mcd43A3) data products
}
\keyword{datasets}
