% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicaragua.R
\docType{data}
\name{nicaragua}
\alias{nicaragua}
\title{Gross primary productivity and evapotranspiration for a farm in Nicaragua}
\format{
A data frame with 360 rows and 4 variables
}
\source{
To be filled in
}
\usage{
data(nicaragua)
}
\description{
\itemize{
  \item date. POSIX date
  \item time. The decimal representation of the date.
  \item product one of Gpp_1km and ET_1km
  \item value value of the remote sensing product
}
}
\keyword{datasets}
