% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_fit_sample.R
\name{periodic_fit_sample}
\alias{periodic_fit_sample}
\title{Compute a baseline and periodic fit to data.}
\usage{
periodic_fit_sample(data_in, regression_formula, percentage)
}
\arguments{
\item{data_in}{a data frame with three columns: date (a POSIX date), time (the decimal time) and the value}

\item{regression_formula}{a formula for the regression that we are computing for the baseline fit}

\item{percentage}{the percentage of the data we are excluding in the fit (via random selection).  Must be between 0 and 1}
}
\value{
A list of two things: the fitted values at a 95% confidence level and the predicted value of data not used in the fitting routine, and an object of class "lm" for analysis.
}
\description{
\code{periodic_fit_sample} Returns fitted and predicted values of periodically repeating data where a certain percentage of the data have been excluded in the fit.
}
\examples{

# To be filled in later
}
