% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_linear_fit.R
\name{piecewise_linear_fit}
\alias{piecewise_linear_fit}
\title{Compute a baseline and piecewise linear fit to data.}
\usage{
piecewise_linear_fit(data_in, regression_formula)
}
\arguments{
\item{data_in}{a data frame with three columns: date (a POSIX date), time (the decimal time) and the value}

\item{regression_formula}{a formula for the regression that we are computing for the baseline fit}
}
\value{
A list of two things: the fitted values at a 95% confidence level and the predicted value, and an object of class "lm" for analysis.
}
\description{
\code{piecewise_linear_fit} Returns fitted and predicted values of periodically repeating data.  This code first subtracts away the residual and does a segmented fit (segmented package)
}
\examples{

}
