% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\docType{data}
\name{synthetic_data}
\alias{synthetic_data}
\title{Synthetic data set for periodic fitting}
\format{A data frame with 1561 rows and 5 variables}
\source{
Zobitz - see the file synthetic-data-process.R in the data-raw folder
}
\usage{
data(synthetic_data)
}
\description{
\itemize{
  \item site. Fluxnet site abbreviation
  \item date. POSIX date.
  \item time. The decimal representation of the date.
  \item product. "GPP" Gross primary productivity
  \item value. value of measurement in previous column
}
}
\keyword{datasets}
