% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_plot.R
\name{taylor_plot}
\alias{taylor_plot}
\title{Make a default Taylor diagram for use}
\source{
Taylor, Karl E. 2001. “Summarizing Multiple Aspects of Model Performance in a Single Diagram.” Journal of Geophysical Research: Atmospheres 106 (D7): 7183–92. https://doi.org/10.1029/2000JD900719.
}
\usage{
taylor_plot(min_R = 0.25, max_R = 1.75, contours = 7, n_lines = 10,
  x_0 = 1, ref_r_min = 0.25, ref_r_max = 2, ref_contours = 8)
}
\arguments{
\item{min_R}{minimum radius of concentric circles}

\item{max_R}{max radius of concentric circles}

\item{contours}{number of concentric circle contours between min_R and max_R}

\item{n_lines}{number of angle contours that measure correlation coeff.}

\item{x_0}{location of data reference on horizontal axis}

\item{ref_r_min}{minimum radius of reference circles from data reference point}

\item{ref_r_max}{maximum radius of reference circles from data reference point}

\item{ref_contours}{number of concentric circle contours between ref_r_min and ref_r_max}
}
\value{
A plot default plotting structure of a circle in the first quadrant.
}
\description{
\code{taylor_plot} Generates a ggplot structure to make a Taylor diagram
}
\examples{

taylor_plot()
}
