% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_plot_process.R
\name{taylor_plot_process}
\alias{taylor_plot_process}
\title{Make a Taylor plot using a list of data}
\usage{
taylor_plot_process(data_results)
}
\arguments{
\item{data_results}{a data frame with eight columns:
\itemize{
\item fit: fitted values
\item lwr: lower confidence value
\item upr: upper confidence value
\item value: the original value
\item site: place (or dataset) utilized
\item model: type of baseline model
\item date: date (a POSIX date),
\item time: (the decimal time)
}

Typically this will be returned by a periodic_fit_process command and then joined together.}
}
\value{
A Taylor diagram
}
\description{
\code{taylor_plot_process} Returns a ggplot for a Taylor diagram using a list of data, which can be then saved.  Prior to this exercise you need to have t
}
\examples{

# To be filled in later
}
