% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_periodic.R
\name{theme_periodic}
\alias{theme_periodic}
\title{Define a custom theme for making ggplots}
\usage{
theme_periodic()
}
\description{
\code{theme_periodic} Sets up defaults for ggplots so I don't need to keep doing these over and over.
}
