% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric_fit_process.R
\name{trigonometric_fit_process}
\alias{trigonometric_fit_process}
\title{Compute a baseline and periodic fit to data.}
\usage{
trigonometric_fit_process(data, formulas)
}
\arguments{
\item{data}{a data frame with three columns: date (a POSIX date), time (the decimal time) and the value}

\item{formulas}{a list of formulas for the regression that we are computing for the baseline fit}
}
\value{
A list of two things: the fitted values at a 95% confidence level and the predicted value, and an object of class "lm" for analysis.
}
\description{
\code{trigonometric_fit_process} Returns a list of fitted and predicted values of periodically repeating data using sine and cosine functions.  This works over a list of data sites efficiently processing multiple locations
}
\examples{

# To be filled in later
}
