% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mauna_loa.R
\docType{data}
\name{mauna_loa}
\alias{mauna_loa}
\title{Mauna Loa carbon dioxide data}
\format{
A data frame with 722 rows and 4 variables
}
\source{
\url{https://www.esrl.noaa.gov/gmd/ccgg/trends/data.html}
I downloaded the data and then modified it in Excel to strip trailing white space, makes it easier to read in.
}
\usage{
data(mauna_loa)
}
\description{
\itemize{
  \item date. POSIX date, but measurement started in 1958.  We have one measurement each month.
  \item time. The decimal representation of the date.
  \item product. = "co2" the measurement
  \item value co2 expressed as a mole fraction in dry air, micromol/mol, abbreviated as ppm
}
}
\keyword{datasets}
